Combine common code between Bukkit and Paper. * That is basically everything, but might be more separated in the future. * Register Commands dynamically, so they work with the `paper-plugin.yml` in place. * Removed WolfyCoreBukkit.getInstance, as that would interfere with WolfyCorePaper. (Use WolfyUtilCore.getInstance instead!) * Replaced WolfyCoreBukkit usages with WolfyUtilCore
