+++ core/src/main/java/com/wolfyscript/utilities/bukkit/commands/InfoCommand.java
@@ -18,7 +18,8 @@
 
 package com.wolfyscript.utilities.bukkit.commands;
 
-import com.wolfyscript.utilities.bukkit.WolfyCoreBukkit;
+import java.util.List;
+import me.wolfyscript.utilities.api.WolfyUtilCore;
 import me.wolfyscript.utilities.api.chat.Chat;
 import me.wolfyscript.utilities.util.version.ServerVersion;
 import net.kyori.adventure.text.Component;
@@ -26,37 +27,39 @@ import net.kyori.adventure.text.format.NamedTextColor;
 import net.kyori.adventure.text.format.TextDecoration;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
-import org.bukkit.command.TabExecutor;
+import org.bukkit.command.PluginIdentifiableCommand;
 import org.bukkit.entity.Player;
+import org.bukkit.plugin.Plugin;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.List;
+public class InfoCommand extends Command implements PluginIdentifiableCommand {
 
-public class InfoCommand implements TabExecutor {
+    private final WolfyUtilCore core;
 
-    private final WolfyCoreBukkit plugin;
+    public InfoCommand(WolfyUtilCore core) {
+        super("wolfyutils");
+        this.core = core;
+        setDescription("Displays info about the plugin version, etc.");
+    }
 
-    public InfoCommand(WolfyCoreBukkit plugin) {
-        this.plugin = plugin;
+    @NotNull
+    @Override
+    public Plugin getPlugin() {
+        return core;
     }
 
     @Override
-    public boolean onCommand(@NotNull CommandSender sender, @NotNull Command command, @NotNull String label, @NotNull String[] args) {
+    public boolean execute(@NotNull CommandSender sender, @NotNull String commandLabel, @NotNull String[] args) {
         if (!(sender instanceof Player)) return true;
-        ((Chat) plugin.getWolfyUtils().getChat()).sendMessages((Player) sender, true,
+        ((Chat) core.getWolfyUtils().getChat()).sendMessages((Player) sender, true,
                 Component.text("——————— ", NamedTextColor.GRAY).append(Component.text("WolfyUtilities", NamedTextColor.AQUA, TextDecoration.BOLD)).append(Component.text(" ———————")),
                 Component.empty(),
-                Component.text("Author: ", NamedTextColor.GRAY).append(Component.text(String.join(", ", plugin.getDescription().getAuthors()), null, TextDecoration.BOLD)),
+                Component.text("Author: ", NamedTextColor.GRAY).append(Component.text(String.join(", ", core.getDescription().getAuthors()), null, TextDecoration.BOLD)),
                 Component.empty(),
                 Component.text("Version: ", NamedTextColor.GRAY).append(Component.text(ServerVersion.getWUVersion().getVersion(), null, TextDecoration.BOLD)),
                 Component.text("———————————————————————", NamedTextColor.GRAY)
         );
         return true;
-    }
-
-    @Override
-    public @Nullable List<String> onTabComplete(@NotNull CommandSender sender, @NotNull Command command, @NotNull String alias, @NotNull String[] args) {
-        return null;
     }
 }
